requirements.txt

altair
annotated-types
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
distro
easyocr
filelock
filetype
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
imageio
Jinja2
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-core
langchain-google-genai
langchain-text-splitters
langsmith
lazy_loader
MarkupSafe
mpmath
narwhals
networkx
ninja
numpy
openai
opencv-python-headless
orjson
packaging
pandas
pillow
pip
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pyclipper
pydantic
pydantic_core
pydeck
pyparsing
python-bidi
python-dateutil
pytz
PyYAML
referencing
requests
requests-toolbelt
rpds-py
rsa
scikit-image
scipy
setuptools
shapely
six
smmap
sniffio
SQLAlchemy
streamlit
sympy
tenacity
tifffile
toml
torch
torchvision
tornado
tqdm
typing_extensions
typing-inspection
tzdata
watchdog
zstandard
